BC-9229 - move userDoFactory back to testing folder
