+++ apps/server/src/modules/authentication/mapper/current-user.mapper.spec.ts
@@ -1,7 +1,6 @@
 import { schoolEntityFactory } from '@modules/school/testing';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
-import { userFactory } from '@modules/user/testing';
+import { userDoFactory, userFactory } from '@modules/user/testing';
 import { ValidationError } from '@shared/common/error';
 import { Permission, RoleName } from '@shared/domain/interface';
 import { setupEntities } from '@testing/database';
