+++ apps/server/src/modules/group/uc/group.uc.spec.ts
@@ -6,9 +6,8 @@ import { roleDtoFactory } from '@modules/role/testing';
 import { SchoolService } from '@modules/school/domain';
 import { schoolEntityFactory, schoolFactory } from '@modules/school/testing';
 import { UserService } from '@modules/user';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
-import { userFactory } from '@modules/user/testing';
+import { userDoFactory, userFactory } from '@modules/user/testing';
 import { ForbiddenException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { NotFoundLoggableException } from '@shared/common/loggable-exception';
