+++ apps/server/src/modules/oauth-provider/domain/service/id-token.service.spec.ts
@@ -4,8 +4,8 @@ import { TeamRepo } from '@modules/team/repo';
 import { teamFactory } from '@modules/team/testing';
 import { externalToolFactory } from '@modules/tool/external-tool/testing';
 import { UserService } from '@modules/user/domain';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
+import { userDoFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { setupEntities } from '@testing/database';
 import { pseudonymFactory } from '@testing/factory/domainobject';
