+++ apps/server/src/modules/provisioning/strategy/tsp/tsp.strategy.spec.ts
@@ -2,7 +2,7 @@ import { faker } from '@faker-js/faker';
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { IdTokenExtractionFailureLoggableException } from '@modules/oauth';
 import { schoolFactory } from '@modules/school/testing';
-import { userDoFactory } from '@modules/user/domain/factory';
+import { userDoFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { RoleName } from '@shared/domain/interface';
 import { SystemProvisioningStrategy } from '@shared/domain/interface/system-provisioning.strategy';
