+++ apps/server/src/modules/user-import/uc/user-import.uc.spec.ts
@@ -15,9 +15,8 @@ import { systemEntityFactory, systemFactory } from '@modules/system/testing';
 import { UserService } from '@modules/user';
 import { UserLoginMigrationService, UserMigrationService } from '@modules/user-login-migration';
 import { userLoginMigrationDOFactory } from '@modules/user-login-migration/testing';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
-import { userFactory } from '@modules/user/testing';
+import { userDoFactory, userFactory } from '@modules/user/testing';
 import { BadRequestException, ForbiddenException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
