+++ apps/server/src/modules/user-login-migration/api/uc/user-login-migration.uc.spec.ts
@@ -12,9 +12,8 @@ import { externalUserDtoFactory } from '@modules/provisioning/testing';
 import { schoolEntityFactory } from '@modules/school/testing';
 import { systemEntityFactory } from '@modules/system/testing';
 import { UserService } from '@modules/user';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
-import { userFactory } from '@modules/user/testing';
+import { userDoFactory, userFactory } from '@modules/user/testing';
 import { ForbiddenException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { NotFoundLoggableException } from '@shared/common/loggable-exception';
