+++ apps/server/src/modules/user-login-migration/domain/service/user-migration.service.spec.ts
@@ -3,7 +3,7 @@ import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { Account, AccountService } from '@modules/account';
 import { UserService } from '@modules/user';
-import { userDoFactory } from '@modules/user/domain/factory';
+import { userDoFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { roleFactory } from '@testing/factory/role.factory';
 import { userLoginMigrationDOFactory } from '../../testing';
