--- apps/server/src/modules/user/domain/factory/index.ts
@@ -1,1 +1,0 @@
-export { userDoFactory } from './user-do.factory';
