+++ apps/server/src/modules/user/testing/index.ts
@@ -1,1 +1,2 @@
+export { userDoFactory } from './user-do.factory';
 export { userFactory } from './user.factory';
