+++ apps/server/src/modules/video-conference/service/video-conference.service.spec.ts
@@ -18,9 +18,8 @@ import { roomFactory } from '@modules/room/testing';
 import { TeamRepo } from '@modules/team/repo';
 import { teamFactory, teamUserFactory } from '@modules/team/testing';
 import { UserService } from '@modules/user';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
-import { userFactory } from '@modules/user/testing';
+import { userDoFactory, userFactory } from '@modules/user/testing';
 import { BadRequestException, ForbiddenException } from '@nestjs/common';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
 import { ConfigService } from '@nestjs/config';
