+++ apps/server/src/modules/video-conference/uc/video-conference-deprecated.uc.spec.ts
@@ -9,8 +9,8 @@ import { LegacySchoolService } from '@modules/legacy-school';
 import { TeamEntity, TeamRepo } from '@modules/team/repo';
 import { teamFactory } from '@modules/team/testing';
 import { UserDo, UserService } from '@modules/user';
-import { userDoFactory } from '@modules/user/domain/factory';
 import { User } from '@modules/user/repo';
+import { userDoFactory } from '@modules/user/testing';
 import { BadRequestException, ForbiddenException, InternalServerErrorException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthorizableObject } from '@shared/domain/domain-object';
